+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 - Change initial default size of Gladiator window to the entire screen size
 - Made layout more resilient to shrinkage so that the Navigation area widgets don't get jumbled up
 - Show open project dialog when run as Gladiator app (e.g. Mac app) instead of opening the local directory of the app
+- Extract gladiator_menu_bar custom widget and reuse in open_project shell (to use in app version)
+- Extract progress_shell custom shell
 - Fix issue with disposed? error in save_config
 - Fix issue with first open of a file with an empty brand new gladiator session not showing file content until resizing